+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [8.2.0] - 2022-09-18
+### Added
+* Added config option to enable/disable trailing slashes support.
+* Added new script to make sure first API key is generated after successfully installing Shlink.
+
+### Changed
+* *Nothing*
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* *Nothing*
+
+### Fixed
+* *Nothing*
+
+
 ## [8.1.0] - 2022-08-08
 ### Added
 * *Nothing*